+++ tests/connectivity-adapter/test/testkit/connector/certs.go
@@ -1,4 +1,4 @@
-package testkit
+package connector
 
 import (
 	"crypto/rand"
