+++ tests/connectivity-adapter/test/testkit/connector/client.go
@@ -1,4 +1,4 @@
-package testkit
+package connector
 
 import (
 	"bytes"
